refactor: move code to net/sample and net/reply files
