+++ src/session/subscription.c
@@ -17,7 +17,9 @@
 #include <stddef.h>
 #include <stdint.h>
 
+#include "zenoh-pico/api/types.h"
 #include "zenoh-pico/config.h"
+#include "zenoh-pico/net/sample.h"
 #include "zenoh-pico/protocol/core.h"
 #include "zenoh-pico/protocol/definitions/network.h"
 #include "zenoh-pico/protocol/keyexpr.h"
